+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.0.0 (September 25, 2020)
+* Include status code, HTTP headers along with body in produced message
+* Update dependencies
+* Remove logging of sensitive data
+* Include attachment information in outbound message
+* Use node version 14
+
 ## 1.2.12 (September 25, 2020)
 * Revert changes to url encoding